+++ builder_pkgs/mockito/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Require Dart >= 2.19.0.
 * Require analyzer 5.11.0.
 * Fail to generate mocks of `sealed`/`base`/`final` classes.
+* Add new `thenReturnInOrder` method to mock multiple calls to a single method in order.
 
 ## 5.4.0